Suggestotron: Implement voting as a custom action on the topics controller
Previously, voting was done through the 'create' action of the 'votes' controller,
which works but does not allow a very sensible symmetry between upvoting and
downvoting. When implemented as a 'upvote' action, the logic required to 'downvote'
becomes simpler.

This commit also moves the beefy 'explaination' at the bottom of the
"allow people to vote" page into the individual steps.